requirements.txt
snap-stanford==6.0.0
